WIP: Custom command
Idea is: if you enter something that doesn't exist as an existing
command, it executes it and remembers it as previous. Currently, if
there's no existing script to run, you have to go to your terminal
first, and more importantly lose the fun F3+enter workflow.
It's pretty trivial to write a script but it would be nice to do
F3+enter with an arbitrary command. It could be listed as a separate
"previous" from existing scripts, i.e. in a different section.

A problem arises: a favorite that is a script can be deleted by deleting
the script, but a favorite that is arbitary can leak data somehow. Need
to make sure it's delete-able. How do you implement the x?